-Fix: CVE-2025-40930 - buffer overread in the FP number parser of the legacy decoder
A specially crafted input string could cause the legacy floating point number parser
(inherited from JSON::XS) to read past allocated memory and thus cause a segmentation fault.
At worst, this could have been used for a denial of service attack.

The JSON::SIMD module was not affected in its default configuration, because the legacy decoder
is disabled by default: the decode_json functional interface always uses the simdjson decoder
and for the OO interface the legacy decoder has to be specifically requested with
JSON::SIMD->new->use_simdjson(0). (There is one exception: the simdjson decoder falls back to
legacy number parser, and thus, it was potentially vulnerable, if the module was compiled with
USE_LONG_DOUBLE or USE_QUADMATH. But again, this is not the default and most binary distributions
of the module were built with these options disabled.)